chore(server): move gql fetchers in gql folder
