+++ packages/twenty-server/test/integration/graphql/suites/settings-permissions/workspace-members.integration-spec.ts
@@ -6,7 +6,7 @@ import { updateOneOperationFactory } from 'test/integration/graphql/utils/update
 import { ErrorCode } from 'src/engine/core-modules/graphql/utils/graphql-errors.util';
 import { PermissionsExceptionMessage } from 'src/engine/metadata-modules/permissions/permissions.exception';
 import { WORKSPACE_MEMBER_DATA_SEED_IDS } from 'src/engine/workspace-manager/dev-seeder/data/constants/workspace-member-data-seeds.constant';
-import { makeGraphqlAPIRequest } from 'test/integration/utils/make-graphql-api-request.util';
+import { makeGraphqlAPIRequest } from 'test/integration/graphql/utils/make-graphql-api-request.util';
 
 const WORKSPACE_MEMBER_GQL_FIELDS = `
     id
