+++ packages/twenty-server/test/integration/graphql/utils/create-one-operation.util.ts
@@ -3,9 +3,9 @@ import { PerformMetadataQueryParams } from 'test/integration/types/perform-metad
 import { capitalize } from 'twenty-shared/utils';
 
 import { ObjectRecord } from 'src/engine/api/graphql/workspace-query-builder/interfaces/object-record.interface';
+import { makeGraphqlAPIRequest } from 'test/integration/graphql/utils/make-graphql-api-request.util';
 import { warnIfNoErrorButExpectedToFail } from 'test/integration/metadata/utils/warn-if-no-error-but-expected-to-fail.util';
 import { CommonResponseBody } from 'test/integration/types/common-response-body.type';
-import { makeGraphqlAPIRequest } from 'test/integration/utils/make-graphql-api-request.util';
 
 type CreateOneOperationArgs<T> = PerformMetadataQueryParams<T> & {
   objectMetadataSingularName: string;
@@ -24,7 +24,9 @@ export const createOneOperation = async <T = object>({
     gqlFields,
   });
 
-  const response = await makeGraphqlAPIRequest<Record<string, ObjectRecord>>({ operation: graphqlOperation });
+  const response = await makeGraphqlAPIRequest<Record<string, ObjectRecord>>({
+    operation: graphqlOperation,
+  });
 
   if (expectToFail) {
     warnIfNoErrorButExpectedToFail({
