+++ packages/twenty-server/test/integration/metadata/suites/field-metadata/utils/create-one-field-metadata.util.ts
@@ -1,9 +1,9 @@
+import { makeMetadataAPIRequest } from 'test/integration/graphql/utils/make-metadata-api-request.util';
 import {
   CreateOneFieldFactoryInput,
   createOneFieldMetadataQueryFactory,
 } from 'test/integration/metadata/suites/field-metadata/utils/create-one-field-metadata-query-factory.util';
 import { PerformMetadataQueryParams } from 'test/integration/types/perform-metadata-query.type';
-import { makeMetadataAPIRequest } from 'test/integration/utils/make-metadata-api-request.util';
 
 import { FieldMetadataEntity } from 'src/engine/metadata-modules/field-metadata/field-metadata.entity';
 import { warnIfNoErrorButExpectedToFail } from 'test/integration/metadata/utils/warn-if-no-error-but-expected-to-fail.util';
