+++ packages/twenty-server/test/integration/metadata/suites/object-metadata/utils/update-one-object-metadata.util.ts
@@ -1,11 +1,11 @@
 import { ObjectRecord } from 'src/engine/api/graphql/workspace-query-builder/interfaces/object-record.interface';
+import { makeGraphqlAPIRequest } from 'test/integration/graphql/utils/make-graphql-api-request.util';
 import {
   UpdateOneObjectFactoryInput,
   updateOneObjectMetadataQueryFactory,
 } from 'test/integration/metadata/suites/object-metadata/utils/update-one-object-metadata-query-factory.util';
 import { warnIfNoErrorButExpectedToFail } from 'test/integration/metadata/utils/warn-if-no-error-but-expected-to-fail.util';
 import { PerformMetadataQueryParams } from 'test/integration/types/perform-metadata-query.type';
-import { makeGraphqlAPIRequest } from 'test/integration/utils/make-graphql-api-request.util';
 
 export const updateOneObjectMetadata = async ({
   input,
